Begun refactoring repository
Added poetry as a build tool and started adding click for args
