+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "core-level-spec-tools"
+version = "2.0.0"
+description = "A collection of tools to perform core-level spectroscopy simulations in CASTEP"
+authors = ["Dylan Morgan <dbmorgan98@gmail.com>"]
+license = "GPL 3"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = ">=3.9, <3.13"
+numpy = "^1.26.2"
+matplotlib = "^3.8.2"
+tqdm = "^4.66.1"
+click = "^8.1.7"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
